make event-key type extensible
Typechecking lisps mean that for every new implemented extension, the type
EVENT-KEY in the main body of CLX needs alteration.  Instead, define a
suitable function for use as a SATISFIES checker, and use it in the type and
in the code.

darcs-hash:20070807091111-30503-ef79b948dff4bec054ef61c0b7ff7efcb1790ea3.gz