colormap / cacheing interaction fixes
Formally, we can't cache colormaps of other clients, or the server; this
means that when we look up a colormap, we will usually be creating one.

That means that we can't rely on the only colormap in town being EQL to
the default colormap, which has its visual-info slot already filled in;
instead, when we lookup a colormap from a window, if its visual-info is
not filled in, set it to the window's visual-info.

[ COLORMAP-VISUAL-INFO is exported from XLIB but not documented in the
  CLX manual.  It might be ripe for non-exporting, particularly given
  that I can't find any internet-accessible code which uses it.  This
  hole was noted from its (unnecessary) use in xrender.lisp ]

darcs-hash:20081008095922-dc3a5-0cf9634db65b42a387dbd4d065080c42b0844299.gz