OR-GET accessor accepts only two types.
First one must be enum (either NULL or MEMBER).

Otherwise if the first type didn't match we could have a condition signalled by
a type-check or by other assertions generated by the compiler (deduced from a
declaration).

Solves a problem with setting a gc-clip mask to a pixmap (it recently started
erroring due to compiler enforcing declarations, but previously didn't work as
well). We fix other OR-GET references in the code to have the correct order.

Closes #97 (and numerous McCLIM issues).