Add support for cmucl
This is a pretty minimal set of changes to get clx to work with cmucl.

Change `*def-clx-class-use-defclass*` to be nil for cmucl.  Not sure
why it was set to use classes for `xlib:drawable`, `xlib:window` and
`xlib:pixmap`.  This is required because dbe wants to include
drawable, which doesn't work when drawable is a class and not a
structure.

Implement `open-x-stream` for cmucl.  This is basically a copy from
the original clx version in cmucl.

Finally, a few minor changes because `with-array-data` is now in the
"LISP" package, not the "KERNEL" package.

With these changes, the demos seem to work.  (The demos want to use
open-clx-display, which isn't defined here.  Not sure what the best
solution is for that.)