Feature clasp take 2 (#136)
implementations: add CLASP to supported platforms

- add missing primitives for clasp
- shadow defconstant (for both sbcl and clasp) to accept objects which are not eql

defconstant which shadows cl:defconstant leaves the value as it were before if it is
already bound. this semantic may not be very satisfactory, but it is better than defparameter:
- it preserves eql'ness
- it doesn't error on lists

One notable issue with it is that it doesn't signal an error if we supply something else in defconstant
later, this new defconstant is ignored. Other implementations (which support defconstant on equal
objects) still use cl:defconstant.