Revert "Memorize the bit depth for drawable".
This commit reverts functional changes in
662e60ce4ad88827441ca6ad04b202cb46e29314.

For context see:
- https://github.com/McCLIM/McCLIM/issues/637
- https://github.com/sharplispers/clx/pull/146
- https://github.com/McCLIM/McCLIM/pull/857

In short: XQuartz X11 server has a very unoptimized code path for
probing the bit depth. Usually this operation is fast, so McCLIM did
check for this bit each time. After finding the issue first fix was
here in clx but that raised a dispute which resulted in this reversal
and caching the depth on McCLIM side.