dependent: use SB-THREAD:WITH-RECURSIVE-LOCK's :TIMEOUT in HOLDING-LOCK
The parameter has been supported for a long time and its use is
essential for correctness - the previous code was subtly broken when
timeouts were used.

To see why, consider the following functions and macros and in
particular the timeout parameter:

  (defun process-event (timeout)
    ...
    (event-loop ... timeout ...)
    ...)

  (defmacro event-loop (... timeout ...)
    (with-event-queue (... timeout ...) ; holding lock with timeout
       ...
       (wait-for-event ...)
       ...))

  (defmacro with-event-queue (... timeout ... body)
     (holding-lock (... timeout ...)
        ...
        body))

  (defun wait-for-event (...)
     ...
     (read-input ...)
     ...)

  (defun read-input (...)
    ...
    (buffer-force-output ...)
    ...)

  (defun buffer-force-output (buffer)
    ...
    (buffer-flush buffer)
    ...)

  (defun buffer-flush (buffer)
     ...
     (buffer-write ...)
     ;; HERE
     (setf (buffer-boffset buffer) 0))

Now consider a client calling PROCESS-EVENT with a non-NIL
timeout. Under the lock held with the specified timeout, a call chain
reaching BUFFER-FORCE-FLUSH can be occur. It is possible for the
specified timeout to expire when execution is at the program point
marked HERE.

When HOLDING-LOCK is implemented using SB-EXT:WITH-TIMEOUT, the
timeout expiration can result in an asynchronous unwind (since the
expiration condition is delivered asynchronously) right at the HERE
program point. This means the buffer contents has already been written
to the output stream but control is transferred before the buffer
offset is reset to 0. This leads to the buffered contents being
written to the output stream a second time when the buffer is next
flushed.