xrender: ensure that the clip mask is always updated
Previously the clip-mask and the server value were compared with EQL when the
picture was synchronized, and the clip-mask was assigned to the server-value.
This approach breaks when the client (preasumbly to avoid consing) modifies
the vector directly. In that case the new value and the server value are EQ
while the clip still needs to be updated.

To avoid this problem when the clip-mask is a sequence, then we ensure that it
is a distinct vector.

Fixes #201.