Demo: Handle "clock" window destroy
When a clock window is destroyed, due to the asynchronous nature of
calls to X11, the rendering code will keep trying to draw objects to the
screen. But since the window is already destroyed, such calls result in
errors like this:

> DRAWABLE-ERROR in current request  Code 14.0 [GetGeometry] ID #x3C00002

The patch fixes the issue by ignoring such errors.

It also stops future rendering once it detects that the window has been
destroyed.

Issue: #186