Demo: Handle "bouncing balls" window destroy
When a bouncing balls window is destroyed, due to the asynchronous
nature of calls to X11, the rendering code will keep trying to draw
objects to the screen. But since the window is already destroyed, such
calls result in errors like this:

> Asynchronous DRAWABLE-ERROR in request 162 (last request was 164) Code
> 63.0 [CopyPlane] ID #x3C00002

The patch fixes the issue by ignoring such errors.

It also stops future rendering once it detects that the window has been
destroyed.

The window name has also been fixed.

Issue: #186